split serial column
